[ALPS05245064] p2p: Fix type mismatch
Fix type mismatch

MTK-Commit-Id: b5fa27ac582e1574cc09b4dd4a801fa05c00a08b

CR-Id: ALPS05245064
Feature: [Module] Wi-Fi Driver

+++ os/linux/gl_cfg80211.c
@@ -6096,7 +6096,7 @@ int mtk_uninit_sta_role(struct ADAPTER *prAdapter,
 int mtk_init_ap_role(struct GLUE_INFO *prGlueInfo,
 		     struct net_device *ndev)
 {
-	int u4Idx = 0;
+	uint8_t u4Idx = 0;
 	struct ADAPTER *prAdapter = prGlueInfo->prAdapter;
 
 	for (u4Idx = 0; u4Idx < KAL_P2P_NUM; u4Idx++) {
